+++ prefect_gitlab/__init__.py
@@ -1,4 +1,6 @@
 from . import _version
-from .filesystems import GitLab  # noqa
+from .credentials import GitLabCredentials
+from .repositories import GitLabRepository
 
 __version__ = _version.get_versions()["version"]
+__all__ = ["GitLabRepository", "GitLabCredentials"]