+++ prefect_gitlab/repositories.py
@@ -4,15 +4,15 @@ Integrations with GitLab.
 The `GitLab` class in this collection is a storage block that lets Prefect agents
 pull Prefect flow code from GitLab repositories.
 
-The `GitLab` block is ideally configured via the Prefect UI, but can also be used 
+The `GitLab` block is ideally configured via the Prefect UI, but can also be used
 in Python as the following examples demonstrate.
 
 Examples:
 ```python
-    from prefect_gitlab.filesystems import GitLab
+    from prefect_gitlab.repositories import GitLabRepository
 
     # public GitLab repository
-    public_gitlab_block = GitLab(
+    public_gitlab_block = GitLabRepository(
         name="my-gitlab-block",
         repository="https://gitlab.com/testing/my-repository.git"
     )
@@ -21,7 +21,7 @@ Examples:
 
 
     # specific branch or tag of a GitLab repository
-    branch_gitlab_block = GitLab(
+    branch_gitlab_block = GitLabRepository(
         name="my-gitlab-block",
         reference="branch-or-tag-name"
         repository="https://gitlab.com/testing/my-repository.git"
@@ -31,7 +31,7 @@ Examples:
 
 
     # private GitLab repository
-    private_gitlab_block = GitLab(
+    private_gitlab_block = GitLabRepository(
         name="my-private-gitlab-block",
         repository="https://gitlab.com/testing/my-repository.git",
         access_token="YOUR_GITLAB_PERSONAL_ACCESS_TOKEN"
@@ -51,19 +51,25 @@ from prefect.exceptions import InvalidRepositoryURLError
 from prefect.filesystems import ReadableDeploymentStorage
 from prefect.utilities.asyncutils import sync_compatible
 from prefect.utilities.processutils import run_process
-from pydantic import Field, HttpUrl, SecretStr, validator
+from pydantic import Field, HttpUrl, validator
+
+from prefect_gitlab.credentials import GitLabCredentials
 
 
-class GitLab(ReadableDeploymentStorage):
+class GitLabRepository(ReadableDeploymentStorage):
     """
     Interact with files stored in GitLab repositories.
+
+    An accessible installation of git is required for this block to function
+    properly.
     """
 
-    _block_type_name = "GitLab"
+    _block_type_name = "GitLab Repository"
     _logo_url = HttpUrl(
         url="https://images.ctfassets.net/gm98wzqotmnx/55edIimT4g9gbjhkh5a3Sp/dfdb9391d8f45c2e93e72e3a4d350771/gitlab-logo-500.png?h=250",  # noqa
         scheme="https",
     )
+    _description = "Interact with files stored in GitLab repositories."
 
     repository: str = Field(
         default=...,
@@ -75,13 +81,13 @@ class GitLab(ReadableDeploymentStorage):
         default=None,
         description="An optional reference to pin to; can be a branch name or tag.",
     )
-    access_token: Optional[SecretStr] = Field(
-        name="Personal Access Token",
+    credentials: Optional[GitLabCredentials] = Field(
         default=None,
-        description="A GitLab Personal Access Token (PAT) with repo scope.",
+        description="An optional GitLab Credentials block for authenticating with "
+        "private GitLab repos.",
     )
 
-    @validator("access_token")
+    @validator("credentials")
     def _ensure_credentials_go_with_https(cls, v: str, values: dict) -> str:
         """Ensure that credentials are not provided with 'SSH' formatted GitLub URLs.
         Note: validates `access_token` specifically so that it only fires when
@@ -91,7 +97,7 @@ class GitLab(ReadableDeploymentStorage):
             if urllib.parse.urlparse(values["repository"]).scheme != "https":
                 raise InvalidRepositoryURLError(
                     (
-                        "Crendentials can only be used with GitHub repositories "
+                        "Credentials can only be used with GitLab repositories "
                         "using the 'HTTPS' format. You must either remove the "
                         "credential if you wish to use the 'SSH' format and are not "
                         "using a private repository, or you must change the repository "
@@ -103,12 +109,12 @@ class GitLab(ReadableDeploymentStorage):
 
     def _create_repo_url(self) -> str:
         """Format the URL provided to the `git clone` command.
-        For private repos: https://<oauth-key>@github.com/<username>/<repo>.git
+        For private repos: https://<oauth-key>@gitlab.com/<username>/<repo>.git
         All other repos should be the same as `self.repository`.
         """
         url_components = urllib.parse.urlparse(self.repository)
-        if url_components.scheme == "https" and self.access_token is not None:
-            token = self.access_token.get_secret_value()
+        if url_components.scheme == "https" and self.credentials is not None:
+            token = self.credentials.token.get_secret_value()
             updated_components = url_components._replace(
                 netloc=f"oauth2:{token}@{url_components.netloc}"
             )
@@ -122,7 +128,7 @@ class GitLab(ReadableDeploymentStorage):
     def _get_paths(
         dst_dir: Union[str, None], src_dir: str, sub_directory: Optional[str]
     ) -> Tuple[str, str]:
-        """Returns the fully formed paths for GitHubRepository contents in the form
+        """Returns the fully formed paths for GitLabRepository contents in the form
         (content_source, content_destination).
         """
         if dst_dir is None:
@@ -143,7 +149,7 @@ class GitLab(ReadableDeploymentStorage):
         self, from_path: Optional[str] = None, local_path: Optional[str] = None
     ) -> None:
         """
-        Clones a GitHub project specified in `from_path` to the provided `local_path`;
+        Clones a GitLab project specified in `from_path` to the provided `local_path`;
         defaults to cloning the repository reference configured on the Block to the
         present working directory.
         Args: